requirements.txt
azure-functions
Django
django-storages[azure]
Pillow
pymysql
